Move settings.js to shell directory.
