+++ src/core/App.vue
@@ -19,7 +19,7 @@ import { loadFieldModules, updateFieldModules } from './field-modules';
 import bundles, { loadConfigDocs, updateConfigDocs } from './store/configDocuments';
 import { alert } from './store/alert';
 import profile, { loadProfile, updateProfile } from './store/profile';
-import settings from './store/settings';
+import settings from './shell/settings';
 
 export default {
   name: 'App',
