+++ src/core/shell/settings.js
